requirement.txt
absl-py==0.9.0
alembic==1.4.3
amqp==2.6.0
aniso8601==7.0.0
anitopy==2.0.0
asgiref==3.2.3
astor==0.8.1
astroid==2.3.3
asyncpg==0.21.0
Babel==2.8.0
bcrypt==3.1.7
Beaker==1.11.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
blinker==1.4
boto==2.49.0
boto3==1.13.8
botocore==1.16.8
bottle==0.12.18
bottle-auth==0.3.3
bottle-beaker==0.1.3
bottle-mongo==0.3.0
bottle-websocket==0.2.9
braintree==4.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chatterbot-corpus==1.2.0
Click==7.0
colorama==0.4.3
configparser==5.0.0
crypto==1.4.1
cryptography==2.8
cssselect2==0.3.0
cycler==0.10.0
databases==0.4.1
decorator==4.4.2
decouple==0.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==1.11.29
django-address==0.2.1
django-admin-chat==0.6
django-admin-views==0.8.0
django-allauth==0.41.0
django-appconf==1.0.4
django-auth-role==0.4.0
django-autoslug==1.9.8
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-channels==0.7.0
django-compressor==2.4
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.0
django-datatables-view==1.19.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extra-views==0.13.0
django-filter==2.2.0
django-google-maps==0.12.1
django-haystack==3.0b2
django-heroku==0.3.1
django-import-export==2.1.0
django-mapbox-location-field==1.4.0
django-modelcluster==5.0.2
django-oscar==2.1
django-oscar-wagtail==0.2.0
django-pandas==0.6.1
django-paypal==1.0.0
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-ses==0.8.14
django-storages==1.9.1
django-suit==0.2.28
django-tables2==2.2.1
django-taggit==1.3.0
django-tinymce==3.0.2
django-treebeard==4.3.1
django-twilio==0.13.0.post1
django-weasyprint==1.0.1
django-webcam-photoupload==0.3.3
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-sav==0.1.0
dnspython==1.16.0
docutils==0.15.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.1.1
fastapi==0.61.2
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
flask-core==2.9.0
Flask-Cors==3.0.8
Flask-Dropzone==1.5.0
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-mongoengine==0.9.5
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
gcloud==0.17.0
geographiclib==1.50
geopy==2.0.0
gevent==20.6.2
gevent-websocket==0.10.1
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==0.4.16
grpcio==1.30.0
h11==0.11.0
h5py==2.10.0
html5lib==1.0.1
httplib2==0.17.0
hupper==1.10.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
iso8601==0.1.13
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
jws==0.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==2.0.0rc3
kivy-deps.angle==0.2.0
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-examples==2.0.0rc3
Kivy-Garden==0.1.4
kiwisolver==1.1.0
kombu==4.6.11
l18n==2018.5
lazy-object-proxy==1.4.3
llvmlite==0.34.0
Mako==1.1.3
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mongoengine==0.20.0
moviepy==1.0.1
mpmath==1.1.0
mysql-connector-python==8.0.20
Naked==0.1.31
nltk==3.5
numba==0.51.1
numpy==1.18.1
oauth2client==3.0.0
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.1
passlib==1.7.2
PasteDeploy==2.1.0
peewee==3.13.3
phonenumbers==8.12.3
Pillow==7.0.0
plaster==1.0
plaster-pastedeploy==0.7
proglog==0.1.9
promise==2.3
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pycryptodomex==3.9.7
pydantic==1.7.2
pygame==1.9.6
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pymunk==5.6.0
pyparsing==2.4.6
Pyphen==0.9.5
pypiwin32==223
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyramid==1.10.4
Pyrebase==3.0.27
Pyrebase4==4.3.0
pyserial==3.5
pystack==1.0.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.12.1
python-jwt==2.0.1
python-socketio==4.5.1
python3-openid==3.1.0
pytz==2020.1
pywin32==228
PyYAML==3.13
rcssmin==1.0.6
redis==3.5.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rjsmin==1.1.0
rosetta==0.3
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.1
serial==0.0.97
shellescape==3.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.13
sqlparse==0.3.1
starlette==0.13.6
stream-django==1.6.1
stream-python==3.2.1
stripe==2.49.0
sympy==1.5.1
tablib==2.0.0
tensorboard==1.12.2
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.0.2
tornado==6.0.4
tqdm==4.43.0
translationstring==1.3
twilio==6.39.0
Unidecode==1.1.1
urllib3==1.25.8
uvicorn==0.12.2
venusian==3.0.0
vine==1.3.0
virtualenv==1.7.1.2
virtualenvwrapper-win==1.2.6
wagtail==2.9
WeasyPrint==51
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.1.0
Willow==1.3
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
zope.deprecation==4.4.0
zope.event==4.4
zope.interface==5.0.2
